+++ dz5/Presenter/Presenter.java
@@ -13,11 +13,11 @@ public class Presenter {
     }
 
     public void buttonClick(){
-        int a = view.getValue("a: ");
-        int b = view.getValue("b: ");
+        double a = view.getValue("a: ");
+        double b = view.getValue("b: ");
         model.setX(a);
         model.setY(b);
-        int result = model.result();
+        double result = model.result();
         view.print(result, "Result: ");
     }
 }
